Refactor announcement and department handling, add DTOs
Replaced Announcement type with AnnouncementDto and updated frontend and backend to use DTOs for announcements, departments, lecturers, modules, and enrollments. Refactored frontend forms and services to use string IDs and improved department/lecturer selection logic. Migrated backend REST resources to controllers, added mappers and DTOs, and updated endpoints to return DTOs. Improved error handling and consistency in entity relationships.

+++ university/src/main/java/com/university/entity/Admin.java
@@ -13,7 +13,7 @@ public class Admin {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer adminId;
+    private Long adminId;
 
     @Column(length = 10, unique = true)
     private String username;
@@ -21,11 +21,11 @@ public class Admin {
     @Column(length = 25)
     private String email;
 
-    public Integer getAdminId() {
+    public Long getAdminId() {
         return adminId;
     }
 
-    public void setAdminId(Integer adminId) {
+    public void setAdminId(Long adminId) {
         this.adminId = adminId;
     }
